+++ packages/vest-isolate/src/runtime.ts
@@ -1,0 +1,1 @@
+export function createRuntime() {}
